# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API do Dicionário",
        "description": "Documentação da API do Dicionário. Fullstack Challenge 🏅 - Dictionary",
        "contact": {
            "email": "suporte@dicionario.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/signin": {
            "post": {
                "tags": [
                    "Gerenciamento de Usuário"
                ],
                "summary": "Realiza a autenticação do usuário.",
                "operationId": "1c84485bee0f29a5c9023c34192496cc",
                "responses": {
                    "200": {
                        "description": "Login successful!"
                    },
                    "500": {
                        "description": "An error occurred, try again!"
                    },
                    "400": {
                        "description": "Invalid credentials!, Validation error!"
                    }
                }
            }
        },
        "/api/auth/signup": {
            "post": {
                "tags": [
                    "Gerenciamento de Usuário"
                ],
                "summary": "Realiza o registro do usuário.",
                "operationId": "33ba43f7bdb523df1786b8e3aa290d2b",
                "responses": {
                    "200": {
                        "description": "Successfully registered!"
                    },
                    "500": {
                        "description": "An error occurred, try again!"
                    },
                    "400": {
                        "description": "Error in registration!"
                    }
                }
            }
        },
        "/api/auth/forgotPassword": {
            "post": {
                "tags": [
                    "Gerenciamento de Usuário"
                ],
                "summary": "Realiza o envio de uma senha aleatória via email para o usuário que esqueceu sua chave de acesso.",
                "operationId": "b183a00fd7c167a2e21daa7da3aca6a2",
                "responses": {
                    "200": {
                        "description": "Email sent successfully!"
                    },
                    "500": {
                        "description": "An error occurred, try again!"
                    },
                    "400": {
                        "description": "Error sending email!, Invalid Name or Email"
                    }
                }
            }
        },
...
        "/api/updateRecord/{id_user}": {
            "put": {
                "tags": [
                    "Gerenciamento de Usuário"
                ],
                "summary": "Realiza a atualização de dados cadastrais do usuário registrado.",
                "operationId": "b9b01dcab28bc3e1d965775bddb4b0f8",
                "responses": {
                    "200": {
                        "description": "Updated successfully!"
                    },
                    "500": {
                        "description": "An error occurred, try again!"
                    },
                    "400": {
                        "description": "Undefined User!, Error when updating!"
                    },
                    "403": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/api/logoutUser/{id_user}": {
            "post": {
                "tags": [
                    "Gerenciamento de Usuário"
                ],
                "summary": "Realiza o logout do usuário atual autenticado",
                "operationId": "f79a61a220a01b882d98413f7d0182d8",
                "responses": {
                    "200": {
                        "description": "Logout completed successfully!"
                    },
                    "500": {
                        "description": "An error occurred, try again!"
                    },
                    "400": {
                        "description": "Undefined User!"
                    },
                    "403": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
        "/api/deleteRecord/{id_user}": {
            "delete": {
                "tags": [
                    "Gerenciamento de Usuário"
                ],
                "summary": "Realiza a exclusão do usuário selecionado do banco de dados",
                "operationId": "684913de67baee7289e0a4a76f32f6c4",
                "responses": {
                    "200": {
                        "description": "Successfully deleted!"
                    },
                    "500": {
                        "description": "An error occurred, try again!"
                    },
                    "400": {
                        "description": "Undefined User!"
                    },
                    "403": {
                        "description": "Unauthorized access."
                    }
                }
            }
        },
...
    },
...
}